requirements.txt
fastapi==0.115.14
uvicorn==0.34.3
motor==3.7.1
pydantic==2.11.7
pydantic_core==2.33.2
annotated-types==0.7.0
httpx==0.28.1
httpcore==1.0.9
anyio==4.9.0
python-dotenv==1.1.1
python-multipart==0.0.20
certifi==2025.6.15
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1